+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Bugfix for checkoutRequest with error code 1053
 - Bugfix class name
+- Allow shipping methods who don't have a delivery date
 
 ## [1.0.4] - 2017-04-24
 ### Changed